requirements.txt
amqp==5.1.1
anytree==2.9.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astroid==2.14.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
beautifulsoup4==4.11.1
billiard==4.1.0
celery==5.3.1
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.30.0
crispy-bootstrap5==0.7
cron-descriptor==1.4.0
cryptography==38.0.4
decouple==0.0.7
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.8
dj-database-url==1.0.0
Django==4.2.3
django-active-link==0.1.8
django-active-users==0.4.0
django-admin-csvexport==2.2
django-admin-data-from-external-service==0.1.2
django-admin-datta==1.0.7
django-allauth==0.51.0
django-bootstrap-icons==0.8.1
django-bootstrap-v5==1.0.11
django-bootstrap5==22.1
django-braces==1.15.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.1
django-countries==7.5.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-decouple==2.1
django-environ==0.8.1
django-facebook==6.0.3
django-filter==22.1
django-heroku==0.3.1
django-ipware==4.0.2
django-modeltree==0.4
django-redis==5.2.0
django-sslserver==0.22
django-timezone-field==5.1
djangorestframework==3.14.0
docker==6.0.0
docopt==0.6.2
ecdsa==0.18.0
engineering-notation==0.6.0
exceptiongroup==1.1.1
ffmpeg==1.4
filelock==3.13.1
Flask==2.2.3
future==0.18.2
gunicorn==20.1.0
idna==3.3
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonfield==3.1.0
kombu==5.3.1
lazy-object-proxy==1.9.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.14.1
marshmallow-jsonapi==0.24.0
mccabe==0.7.0
misaka==2.1.1
numpy==1.23.5
oauthlib==3.2.2
oic==1.5.0
packaging==21.3
pandas==1.5.0
paystack==1.5.0
Pillow==9.2.0
platformdirs==4.1.0
pluggy==1.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.16.0
pydub==0.25.1
pyjwkest==1.4.2
PyJWT==2.6.0
pylint==2.16.2
pyparsing==3.0.9
pypaystack2==1.1.3
pytest==7.1.3
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.3
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stripe==5.2.0
tk==0.1.0
tk-tools==0.16.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2023.3
Unidecode==1.3.6
Unipath==1.1
urllib3==1.26.11
vine==5.0.0
virtualenv==20.25.0
wcwidth==0.2.6
websocket-client==1.3.3
Werkzeug==2.2.3
whitenoise==6.2.0
wrapt==1.14.1
